add title card compt, import to LMap
